Docker starter project for CouchDb and Makeswap
Because:
- Mkswap script to create a swapfile in each Ubuntu droplet
- CouchDB is a NoSQL database with RESTful APIs

As the droplet is based on a standard plan with the least memory, it is
necessary to create a swapfile to ensure that we have sufficient
memory to run our web applications.

CouchDb container is deployed to a droplet, which provides a backend
NoSQL database with RESTful APIs.

For example, Inkdrop uses couchDb for cloud sync.

Another example, we can use couchDb as a backend store for our stocks
financial data.

Also, we can use couchDb to broadcast a master MT4 account trades to
multiple slave MT4 accounts.

These changes should resolve:
- automate make swapfile on each Ubuntu droplet
- CouchDB container on each Ubuntu droplet
- Security of couchDb is admin and password

However, it does not resolve:
- reverse proxy
- SSL for couchDb