Docker starter project is production-ready for DO.
This commit is production ready for a DigitalOcean account. The modules
in production are teedy and couchdb, each with its own droplet in their
separate projects. Each module consists of three files: resource, output
and variable files.

However, this commit has neither Nginx reverse proxy nor SSL, which
limits the use of the droplets.

This commit resolves:
- provisioning of files, such as swap, docker-compose, and data files
- restore of data files to droplets
- execution of make swapfile
- execution of docker-compose up

This commit does not resolve:
- provisioning of providers
- backup of data files from droplets