Add new DO project and module Jitsi to Terraform
Updates:
- add new module Jitsi, with own resources, variables, and outputs files
- automatic assign Jitsi (previously Teedy) droplet IP to custom domain

The module Jitsi, is based on a pre-configured image with Jitsi and NGINX,
by DigitalOcean. It is recommended to run this image on a CPU-optimized
droplet with at least 4GB of RAM.

The image provides scripts for post-droplet setup with HTTPS for secure
video conferencing. Hence, a custom domain is required, as it is not
possible to enable HTTPS with just an IP.

After the Jitsi droplet is ready, access the console via SSH. Navigate
to the /root/ folder, then execute the following bash scripts.
1. Configure your video conference and enter the domain name by typing:
  $ ./01_videoconf.sh
2. Configure your Let's Encrypt SSL and enter your email by typing:
  $ ./02_https.sh

This commit does not resolve:
- provision execution of post-droplet scripts for Jitsi module
- provision providers, such as AWS
- provision backup of files from droplets